+++ packages/gamut-kit/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.6.22](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.21...@codecademy/gamut-kit@0.6.22) (2021-11-30)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.6.21](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.20...@codecademy/gamut-kit@0.6.21) (2021-11-29)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
 ### [0.6.20](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.19...@codecademy/gamut-kit@0.6.20) (2021-11-23)
 
 **Note:** Version bump only for package @codecademy/gamut-kit